requirements.txt
requests>=2.31.0
websockets>=11.0.3
pydub>=0.25.1
simpleaudio>=1.0.4
python-dotenv>=1.0.0 